Created template class for JoyDPad and VDPad xml (backup)
