+++ src/setjoystick.h
@@ -30,11 +30,6 @@ class JoyButton;
 class JoyDPad;
 class JoyControlStick;
 class VDPad;
-class QXmlStreamReader;
-class QXmlStreamWriter;
-class JoyDPadXml;
-class JoyAxisXml;
-class JoyButtonXml;
 
 class SetJoystick : public QObject
 {
@@ -88,11 +83,11 @@ public:
     void removeAllBtnFromQueue();
     int getCountBtnInList(QString partialName);
 
-    virtual void readConfig(QXmlStreamReader *xml);
-    virtual void writeConfig(QXmlStreamWriter *xml);
+    bool isSetEmpty();
+
 
 protected:
-    bool isSetEmpty();
+
     void deleteButtons(); // SetButton class
     void deleteAxes(); // SetAxis class
     void deleteHats(); // SetHat class
@@ -184,10 +179,6 @@ private:
     int m_index;
     InputDevice *m_device;
     QString m_name;
-
-    JoyDPadXml* joydpadXml;
-    JoyAxisXml* joyAxisXml;
-    JoyButtonXml* joyButtonXml;
 };
 
 Q_DECLARE_METATYPE(SetJoystick*)
